+++ rmk/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Optimize nRF BLE power consumption, now the idle current is decreased to about 20uA
 - Added new `wm` "With Modifier" macro to support basic keycodes with modifiers active
 - Direct pin mod. Including `DirectPinMatrix`, `run_rmk_direct_pin` functions etc.
+- Added pin active level parameter `low_active` to direct pin.
 
 ### Changed